add stream schema
